Update OphysSession.m
Fix type mismatch

+++ +bot/+behavior/+item/OphysSession.m
@@ -145,7 +145,7 @@ classdef OphysSession < bot.behavior.item.internal.abstract.Item & ...
     
             % Get cells for experiments of session
             cells = bot.behavior.listCells();
-            obj.Cells = cells(ismember(cells.ophys_experiment_id, uint64(experimentIds)), :);
+            obj.Cells = cells(ismember(cells.ophys_experiment_id, experimentIds), :);
             obj.Cells = sortrows(obj.Cells, 'cell_roi_id');
         end
